requirements.txt
anyascii==0.3.2
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
dj-database-url==1.2.0
Django==4.1.7
django-on-heroku==1.1.2
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
psycopg2-binary==2.9.5
pyahocorasick==2.0.0
PyJWT==2.6.0
python-decouple==3.8
pytz==2022.7.1
rebrand==0.0.12
requests==2.28.2
sqlparse==0.4.3
textsearch==0.0.24
twilio==7.16.5
urllib3==1.26.15
whitenoise==6.4.0
